getActiveTriggers van WebSockets naar Long Polling gezet, WebSockets naar TitleBarContainer verplaatst
