+++ src/components/Triggers/TriggerCreateCategory.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import TriggerCreateCategory from './TriggerCreateCategory';
 import { mount } from 'enzyme';
 
-describe('(Container) TriggerCreateCategory', () => {
+describe('(Component) TriggerCreateCategory', () => {
   it('Renders a TriggerCreateCategory', () => {
     const _component = mount(<TriggerCreateCategory icon='star' />);
     expect(_component.type()).to.eql(TriggerCreateCategory);
